commands: Create resumeStateIfExploreNotebook function
